Modified OWObject_Destroy and fixed a value instantiation

+++ src/object.c
@@ -15,7 +15,7 @@ OWObject_t* _OWObject_Construct(size_t size, OWID type, OWDestroyCallback_t* des
   }
 
   this->super = super;
-  this->reference_count += 1;
+  this->reference_count = 1;
   this->destroy_callback = destroy_callback;
   this->type = type;
   return this;
@@ -23,19 +23,11 @@ OWObject_t* _OWObject_Construct(size_t size, OWID type, OWDestroyCallback_t* des
 
 
 void OWObject_Destroy(OWObject_t* this) {
-  OWObject_t* prev_this = this;
-  while(this != NULL) {
-    if(this->destroy_callback != NULL) {
-      this->destroy_callback(this);
-    }
-    if(this->object != NULL) {
-      free(this->object);
-    }
-
-    this = this->super;
-    free(prev_this);
-    prev_this = this;
-  }
+  OWObject_t* const super = this->super;
+  if(this->destroy_callback != NULL) this->destroy_callback(this);
+  if(this->object != NULL) free(this->object);
+  free(this);
+  if(super != NULL) OWObject_UnRef(super);
 }
 
 OWObject_t* OWObject_FindTypeInClass(OWObject_t* this, OWID type) {
